+++ BirtherDay/Presentation/CreateCoupon/CouponCompleteView.swift
@@ -18,9 +18,6 @@ struct CouponCompleteView: View {
     @State private var isLoading: Bool = false
     
     //TODO: 삭제
-    private let shareManager: ShareManager = ShareManager(
-        message: "사랑하는 길님의 생일쿠폰이 도착했어요.\n쿠폰함을 확인해보세요.", params: ["test1": "value"], shareType: .coupon
-    )
     
     private let shareModalHeight: CGFloat = 195
     
@@ -111,11 +108,8 @@ struct CouponCompleteView: View {
                 .font(.b2)
                 .padding(.top, 16)
             
-            HStack(spacing: 25) {
-                kakaoShareButtonView()
-                moreShareButtonView()
-            }
-            .padding(.top, 24)
+            ShareButtons()
+                .padding(.top, 24)
         }
     }
     
@@ -132,50 +126,6 @@ struct CouponCompleteView: View {
             Text("홈으로")
         }
         .buttonStyle(BDButtonStyle(buttonType: .activate))
-    }
-    
-    func kakaoShareButtonView() -> some View {
-        VStack {
-            Button {
-                shareManager.shareToKakao()
-            } label: {
-                Image("kakaoIcon")
-                    .resizable()
-                    .frame(
-                        width: 57,
-                        height: 57
-                    )
-            }
-            
-            Text("카카오톡")
-                .font(.r1)
-        }
-    }
-    
-    func moreShareButtonView() -> some View {
-        VStack {
-            Button {
-                let shareLinkData = shareManager.getShareLinkData()
-                ShareLink(
-                    item: shareLinkData.photo!
-                    , message: Text(shareLinkData.message)
-                    , preview: SharePreview(
-                        "",
-                        image: shareLinkData.photo!.image
-                    )
-                )
-            } label: {
-                Image("moreIcon")
-                    .resizable()
-                    .frame(
-                        width: 57,
-                        height: 57
-                    )
-            }
-            
-            Text("더보기")
-                .font(.r1)
-        }
     }
 }
